requirements.txt
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
et-xmlfile==1.1.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.0
httplib2==0.22.0
idna==3.7
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.1
six==1.16.0
soupsieve==2.5
StrEnum==0.4.15
tzdata==2024.1
urllib3==2.2.1
